requirements.txt
anthropic>=0.25.0
python-osc>=1.8.0
python-dotenv>=1.0.0
